# backend/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: CERBERUS API
  description: API definition for the CERBERUS implant-management API.
  version: 2.0.0
# TODO Update with ACG stuff
components:
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
...
paths:
...
  /users/user/{username}: 
    get: 
      security:
        - jwt: [Read]
      summary: Gets the details (minus hashed password) of the user, if they exist.
      tags:
        - Users
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Name of the user to find.
      responses:
        "200": # status code
          description: Successfully found user
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string
        "400": # status code
          description: Provided username does not correspond to a user
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string
        "403": # status code
          description: Calling user not logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string
  /users/user/{userId}: 
    delete: 
      security:
        - jwt: [Admin]
      summary: Ensures a user does not exist in the database. Requires admin access. Does not work when auth is not DB-backed.
      tags:
        - Users
      parameters:
        - in: path
          name: userId
          schema:
            type: string
          required: true
          description: Identifier of the user to delete.
      responses:
        "200": # status code
          description: User deleted or did not exist.
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string
        "403": # status code
          description: Calling user not logged in or not an admin.
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string
        "500": # status code
          description: Server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string
  /users/whoami:
    get:
      security:
        - jwt: [Read]
      summary: Checks the token for the user, and returns their details.
      tags:
        - Users
      responses:
        "200": # status code
          description: User is logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: 
                    type: object
                    properties:
                      id: 
                        type: string
                      name:
                        type: string
                      isAdmin:
                        type: boolean
                  errors:
                    type: array
                    items:
                      type: string
        "401": # status code
          description: User is not logged in.
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string
        "500": # status code
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string